+++ fanboi2/tests/test_tasks.py
@@ -67,31 +67,40 @@ class TestAddTopicTask(TaskMixin, ModelMixin, unittest.TestCase):
         self.assertEqual(topic.posts[0].body, 'Hello, world!')
         self.assertEqual(result.result, ('topic', topic.id))
 
-    @unittest.mock.patch('fanboi2.utils.Akismet.spam')
-    def test_add_topic_spam(self, akismet):
+    def test_add_topic_overridden(self):
+        import transaction
         from fanboi2.models import Topic
-        akismet.return_value = True
-        request = {'remote_addr': '127.0.0.1'}
+        request = {'remote_addr': '10.0.1.1'}
         with transaction.manager:
-            board = self._makeBoard(title='Foobar', slug='foobar')
+            self._makeRuleOverride(
+                ip_address='10.0.1.0/24',
+                override={'status': 'open'})
+            board = self._makeBoard(
+                title='Foobar',
+                slug='foobar',
+                status='restricted')
             board_id = board.id  # board is not bound outside transaction!
         result = self._makeOne(request, board_id, 'Foobar', 'Hello, world!')
+        topic = DBSession.query(Topic).first()
+        print(result.result)
         self.assertTrue(result.successful())
-        self.assertEqual(DBSession.query(Topic).count(), 0)
-        self.assertEqual(result.result, ('failure', 'spam_rejected'))
+        self.assertEqual(DBSession.query(Topic).count(), 1)
+        self.assertEqual(DBSession.query(Topic).get(result.get()[1]), topic)
+        self.assertEqual(topic.title, 'Foobar')
+        self.assertEqual(topic.posts[0].body, 'Hello, world!')
+        self.assertEqual(result.result, ('topic', topic.id))
 
-    @unittest.mock.patch('fanboi2.utils.Dnsbl.listed')
-    def test_add_topic_dnsbl(self, dnsbl):
+    def test_add_topic_ban(self):
         from fanboi2.models import Topic
-        dnsbl.return_value = True
-        request = {'remote_addr': '127.0.0.1'}
+        request = {'remote_addr': '10.0.1.1'}
         with transaction.manager:
+            self._makeRuleBan(ip_address='10.0.1.0/24')
             board = self._makeBoard(title='Foobar', slug='foobar')
             board_id = board.id  # board is not bound outside transaction!
         result = self._makeOne(request, board_id, 'Foobar', 'Hello, world!')
         self.assertTrue(result.successful())
         self.assertEqual(DBSession.query(Topic).count(), 0)
-        self.assertEqual(result.result, ('failure', 'dnsbl_rejected'))
+        self.assertEqual(result.result, ('failure', 'ban_rejected'))
 
     def test_add_topic_board_restricted(self):
         from fanboi2.models import Topic
@@ -144,6 +153,32 @@ class TestAddTopicTask(TaskMixin, ModelMixin, unittest.TestCase):
             'status_rejected',
             'archived'))
 
+    @unittest.mock.patch('fanboi2.utils.Akismet.spam')
+    def test_add_topic_spam(self, akismet):
+        from fanboi2.models import Topic
+        akismet.return_value = True
+        request = {'remote_addr': '127.0.0.1'}
+        with transaction.manager:
+            board = self._makeBoard(title='Foobar', slug='foobar')
+            board_id = board.id  # board is not bound outside transaction!
+        result = self._makeOne(request, board_id, 'Foobar', 'Hello, world!')
+        self.assertTrue(result.successful())
+        self.assertEqual(DBSession.query(Topic).count(), 0)
+        self.assertEqual(result.result, ('failure', 'spam_rejected'))
+
+    @unittest.mock.patch('fanboi2.utils.Dnsbl.listed')
+    def test_add_topic_dnsbl(self, dnsbl):
+        from fanboi2.models import Topic
+        dnsbl.return_value = True
+        request = {'remote_addr': '127.0.0.1'}
+        with transaction.manager:
+            board = self._makeBoard(title='Foobar', slug='foobar')
+            board_id = board.id  # board is not bound outside transaction!
+        result = self._makeOne(request, board_id, 'Foobar', 'Hello, world!')
+        self.assertTrue(result.successful())
+        self.assertEqual(DBSession.query(Topic).count(), 0)
+        self.assertEqual(result.result, ('failure', 'dnsbl_rejected'))
+
 
 class TestAddPostTask(TaskMixin, ModelMixin, unittest.TestCase):
 
@@ -168,20 +203,42 @@ class TestAddPostTask(TaskMixin, ModelMixin, unittest.TestCase):
         self.assertEqual(post.bumped, True)
         self.assertEqual(result.result, ('post', post.id))
 
-    @unittest.mock.patch('fanboi2.utils.Akismet.spam')
-    def test_add_post_spam(self, akismet):
+    def test_add_post_overridden(self):
         import transaction
         from fanboi2.models import Post
-        akismet.return_value = True
-        request = {'remote_addr': '127.0.0.1'}
+        request = {'remote_addr': '10.0.1.1'}
         with transaction.manager:
+            self._makeRuleOverride(
+                ip_address='10.0.1.0/24',
+                override={'status': 'open'})
+            board = self._makeBoard(
+                title='Foobar',
+                slug='foobar',
+                status='locked')
+            topic = self._makeTopic(board=board, title='Hello, world!')
+            topic_id = topic.id  # topic is not bound outside transaction!
+        result = self._makeOne(request, topic_id, 'Hi!', True)
+        post = DBSession.query(Post).first()
+        self.assertTrue(result.successful())
+        self.assertEqual(DBSession.query(Post).count(), 1)
+        self.assertEqual(DBSession.query(Post).get(result.get()[1]), post)
+        self.assertEqual(post.body, 'Hi!')
+        self.assertEqual(post.bumped, True)
+        self.assertEqual(result.result, ('post', post.id))
+
+    def test_add_post_ban(self):
+        import transaction
+        from fanboi2.models import Post
+        request = {'remote_addr': '10.0.1.1'}
+        with transaction.manager:
+            self._makeRuleBan(ip_address='10.0.1.0/24')
             board = self._makeBoard(title='Foobar', slug='foobar')
             topic = self._makeTopic(board=board, title='Hello, world!')
             topic_id = topic.id  # topic is not bound outside transaction!
         result = self._makeOne(request, topic_id, 'Hi!', True)
         self.assertTrue(result.successful())
         self.assertEqual(DBSession.query(Post).count(), 0)
-        self.assertEqual(result.result, ('failure', 'spam_rejected'))
+        self.assertEqual(result.result, ('failure', 'ban_rejected'))
 
     def test_add_post_locked(self):
         import transaction
@@ -243,6 +300,21 @@ class TestAddPostTask(TaskMixin, ModelMixin, unittest.TestCase):
             'failure',
             'status_rejected',
             'archived'))
+
+    @unittest.mock.patch('fanboi2.utils.Akismet.spam')
+    def test_add_post_spam(self, akismet):
+        import transaction
+        from fanboi2.models import Post
+        akismet.return_value = True
+        request = {'remote_addr': '127.0.0.1'}
+        with transaction.manager:
+            board = self._makeBoard(title='Foobar', slug='foobar')
+            topic = self._makeTopic(board=board, title='Hello, world!')
+            topic_id = topic.id  # topic is not bound outside transaction!
+        result = self._makeOne(request, topic_id, 'Hi!', True)
+        self.assertTrue(result.successful())
+        self.assertEqual(DBSession.query(Post).count(), 0)
+        self.assertEqual(result.result, ('failure', 'spam_rejected'))
 
     def test_add_post_retry(self):
         import transaction